refactor: move WinningNumber
